Add main server file with basic routing
This server only responds to requests on /.  It expects POST requests
from Slack, so that's where the main logic will be, and it outputs the
name of the service on GET requests so you can know you're at the right
place when you ping it.  Any other request results in an error.