Add basic time-to-emoji converter with rounding
So far this only handles the minutes, it checks whether it's closer to
the previous hour, the next/previous half hour or the next hour and
returns emoji text based on that.

So far it doesn't use the hours from the provided time, it's just hard
coded for 12 o'clock, so that needs to be improved.  I add several test
cases based on manual tests, it might need more though.  It also
involves getting the time from the arguments' text field, parsing that
and returning an error on failure.  Currently that error message still
sucks.